feat(loader): migrate to ts
