+++ packages/loader/package.json
@@ -2,7 +2,7 @@
   "name": "@axa-fr/react-toolkit-loader",
   "version": "1.3.14",
   "main": "dist/index.js",
-  "jsnext:main": "src/index.js",
+  "types": "dist/index.d.ts",
   "private": false,
   "scripts": {
     "prepare": "node ../../scripts/prepare.js"
